+++ src/surfgreen.rs
@@ -10,6 +10,7 @@
 use crate::error::{TbError, Result};
 use crate::{Model};
 use crate::kpoints::gen_kmesh;
+use crate::basis::Dimension;
 use gnuplot::Major;
 use gnuplot::{Auto, AutoOption::Fix, AxesCommon, Custom, Figure, Font, HOT, RAINBOW};
 use ndarray::concatenate;
@@ -80,10 +81,10 @@ impl surf_Green {
     ///
     ///对于非晶格矢量得方向, 需要用 model.make_supercell 先扩胞
     pub fn from_Model(model: &Model, dir: usize, eta: f64, Np: Option<usize>) -> Result<surf_Green> {
-        if dir >= model.dim_r {
+        if dir >= model.dim_r as usize {
             return Err(TbError::InvalidDirection {
                 index: dir,
-                dim: model.dim_r,
+                dim: model.dim_r as usize,
             });
         }
         let mut R_max: usize = 0;
@@ -103,13 +104,13 @@ impl surf_Green {
             None => R_max,
         };
 
-        let mut U = Array2::<f64>::eye(model.dim_r);
+        let mut U = Array2::<f64>::eye(model.dim_r as usize);
         U[[dir, dir]] = R_max as f64;
         let model = model.make_supercell(&U)?;
         let mut ham0 = Array3::<Complex<f64>>::zeros((0, model.nsta(), model.nsta()));
-        let mut hamR0 = Array2::<isize>::zeros((0, model.dim_r));
+        let mut hamR0 = Array2::<isize>::zeros((0, model.dim_r as usize));
         let mut hamR = Array3::<Complex<f64>>::zeros((0, model.nsta(), model.nsta()));
-        let mut hamRR = Array2::<isize>::zeros((0, model.dim_r));
+        let mut hamRR = Array2::<isize>::zeros((0, model.dim_r as usize));
         let use_hamR = model.hamR.rows();
         let use_ham = model.ham.outer_iter();
         for (ham, R) in use_ham.zip(use_hamR) {
@@ -130,7 +131,7 @@ impl surf_Green {
         let new_hamR0 = remove_col(hamR0, dir);
         let new_hamRR = remove_col(hamRR, dir);
         let green: surf_Green = surf_Green {
-            dim_r: model.dim_r - 1,
+            dim_r: model.dim_r as usize - 1,
             norb: model.norb(),
             nsta: model.nsta(),
             natom: model.natom(),
